+++ equ-clt/src/main/java/equ/clt/EquipmentServer.java
@@ -561,11 +561,11 @@ public class EquipmentServer {
         //groupId заданий, находящихся в обработке
         Set<String> currentlyProceededGroups;
         //задания в очереди
-        Map<Integer, TransactionInfo> waitingTaskQueueMap;
+        Map<Long, TransactionInfo> waitingTaskQueueMap;
         //выполняющиеся задания
-        List<Integer> proceededTaskList;
+        List<Long> proceededTaskList;
         //выполненные задания
-        List<Integer> succeededTaskList;
+        List<Long> succeededTaskList;
 
         public TaskPool(EquipmentServerInterface remote, String sidEquipmentServer) {
             this.remote = remote;
@@ -607,8 +607,8 @@ public class EquipmentServer {
             if(minGroupId != null) {
                 currentlyProceededGroups.add(minGroupId);
                 resultTask = new SingleTransactionTask(remote, minGroupId, minClsHandler, new ArrayList<TransactionInfo>(), sidEquipmentServer);
-                Set<Integer> removingTaskSet = new HashSet<>();
-                for (Map.Entry<Integer, TransactionInfo> transactionInfo : waitingTaskQueueMap.entrySet()) {
+                Set<Long> removingTaskSet = new HashSet<>();
+                for (Map.Entry<Long, TransactionInfo> transactionInfo : waitingTaskQueueMap.entrySet()) {
                     if(resultTask.groupId.equals(getTransactionInfoGroupId(transactionInfo.getValue()))) {
                         processTransactionLogger.info(String.format("Task Pool : starting transaction %s", transactionInfo.getValue().id));
                         resultTask.transactionEntry.add(transactionInfo.getValue());
@@ -617,7 +617,7 @@ public class EquipmentServer {
                     }
                 }
                 Collections.sort(resultTask.transactionEntry, COMPARATOR);
-                for(Integer task : removingTaskSet)
+                for(Long task : removingTaskSet)
                     waitingTaskQueueMap.remove(task);
             }
             return resultTask;
@@ -625,7 +625,7 @@ public class EquipmentServer {
 
         //метод, считывающий задания из базы
         synchronized void addTasks(List<TransactionInfo> transactionInfoList) throws Exception {
-            Map<Integer, TransactionInfo> newWaitingTaskQueueMap = new OrderedMap<>();
+            Map<Long, TransactionInfo> newWaitingTaskQueueMap = new OrderedMap<>();
             for(TransactionInfo transaction : transactionInfoList) {
                 if(!succeededTaskList.contains(transaction.id) && !proceededTaskList.contains(transaction.id)) {
                     if(!waitingTaskQueueMap.containsKey(transaction.id))
@@ -644,8 +644,8 @@ public class EquipmentServer {
                 if(transactionEntry.getValue())
                     succeededTaskList.add(transactionEntry.getKey().id);
                 else {
-                    for(Iterator<Map.Entry<Integer, TransactionInfo>> it = waitingTaskQueueMap.entrySet().iterator(); it.hasNext(); ) {
-                        Map.Entry<Integer, TransactionInfo> entry = it.next();
+                    for(Iterator<Map.Entry<Long, TransactionInfo>> it = waitingTaskQueueMap.entrySet().iterator(); it.hasNext(); ) {
+                        Map.Entry<Long, TransactionInfo> entry = it.next();
                         if(groupId != null && groupId.equals(getTransactionInfoGroupId(entry.getValue()))) {
                             it.remove();
                         }
@@ -712,7 +712,7 @@ public class EquipmentServer {
                 }
 
                 try {
-                    Map<Integer, SendTransactionBatch> succeededMachineryInfoMap = clsHandler.sendTransaction(transactionEntry);
+                    Map<Long, SendTransactionBatch> succeededMachineryInfoMap = clsHandler.sendTransaction(transactionEntry);
                     
                     processTransactionLogger.info(String.format("   Sending transaction group %s: confirm to server, count : %s ", groupId, succeededMachineryInfoMap.size()));
 
@@ -772,7 +772,7 @@ public class EquipmentServer {
             return enabledMachineryInfoList.isEmpty() ? machineryInfoList : enabledMachineryInfoList;
         }
 
-        private void errorTransactionReport(Integer idTransactionInfo, Throwable e) {
+        private void errorTransactionReport(Long idTransactionInfo, Throwable e) {
             try {
                 remote.errorTransactionReport(idTransactionInfo, e);
             } catch (Exception ignored) {
@@ -780,7 +780,7 @@ public class EquipmentServer {
             }
         }
 
-        private void succeededTransaction(Integer idTransactionInfo) {
+        private void succeededTransaction(Long idTransactionInfo) {
             try {
                 remote.succeedTransaction(idTransactionInfo, new Timestamp(Calendar.getInstance().getTime().getTime()));
             } catch (Exception ignored) {
